Splits cli into .h and .c; also code cleanup
